Move Gatsby config files to TypeScript
