+++ gatsby/gatsby-ts/gatsby-browser/index.ts
@@ -1,0 +1,2 @@
+export * from './onRouteUpdate';
+export * from './wrapRootElement';
