+++ gatsby/gatsby-ts/gatsby-node/createPages.ts
@@ -1,0 +1,81 @@
+import { CreatePagesArgs, GatsbyNode } from 'gatsby';
+
+interface INode {
+  fileAbsolutePath: string;
+  frontmatter: {
+    description: string;
+    draft: boolean;
+    keywords: string[];
+    title: string;
+  };
+  id: string;
+  // TODO: Properly type parent object
+  parent: any;
+  // TODO: Properly type table of contents object
+  tableOfContents: any;
+}
+
+interface IQueryResult {
+  allMdx: {
+    edges: {
+        node: INode;
+    }[];
+  };
+}
+
+export const createPages: GatsbyNode['createPages'] = async (
+  props: CreatePagesArgs
+) => {
+  const { createPage } = props.actions;
+  const result = await props.graphql<IQueryResult>(`
+    query {
+      allMdx {
+        edges {
+          node {
+            fileAbsolutePath
+            frontmatter {
+              title
+              description
+              keywords
+              draft
+            }
+            id
+            tableOfContents(maxDepth: 3)
+            parent {
+              id
+              ... on File {
+                id
+                name
+                relativeDirectory
+              }
+            }
+          }
+        }
+      }
+    }
+  `);
+
+  if (result.errors) {
+    props.reporter.panicOnBuild('🚨 ERROR: Loading "createPages" query');
+  }
+
+  if (!result.data) {
+    props.reporter.panicOnBuild('🚨 ERROR: Loading "createPages" query');
+    throw new Error('🚨 ERROR: Loading "createPages" query');
+  }
+
+  const posts = result.data.allMdx.edges;
+
+  posts.forEach(({ node }) => {
+    if (
+      process.env.gatsby_executing_command === 'develop'
+      || !node.frontmatter.draft
+    ) {
+      createPage({
+        component: node.fileAbsolutePath,
+        context: node,
+        path: `${node.parent.relativeDirectory}/${node.parent.name}`,
+      });
+    }
+  });
+};
