+++ gatsby/gatsby-ts/gatsby-node/index.ts
@@ -1,0 +1,2 @@
+export * from './createPages';
+export * from './onCreateWebpackConfig';
