+++ gatsby/gatsby-ts/gatsby-ssr/index.ts
@@ -1,0 +1,2 @@
+export * from './replaceRenderer';
+export * from './wrapRootElement';
