+++ gatsby/gatsby-ts/types/remark-slug/index.d.ts
@@ -1,0 +1,4 @@
+declare module 'remark-slug' {
+  const content: any;
+  export default content;
+}
