+++ gatsby/package.json
@@ -4,9 +4,13 @@
     "url": "https://github.com/maxmind/dev-site/issues"
   },
   "dependencies": {
+    "@babel/preset-env": "^7.10.3",
+    "@babel/preset-react": "^7.10.1",
     "@mdx-js/mdx": "^1.5.8",
     "@mdx-js/react": "^1.5.8",
+    "@types/cheerio": "^0.22.18",
     "@types/classnames": "^2.2.10",
+    "@types/mdx-js__react": "^1.5.2",
     "@types/react": "^16.9.33",
     "@types/react-dom": "^16.9.7",
     "@types/react-helmet": "^5.0.15",
@@ -40,6 +44,7 @@
     "react-helmet": "^5.1.3",
     "react-icons": "^3.10.0",
     "remark-slug": "^6.0.0",
+    "ts-node": "^8.10.2",
     "typeface-roboto": "^0.0.75",
     "typeface-source-code-pro": "^1.1.3",
     "typescript": "^3.8.3"
