+++ gatsby/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     "esModuleInterop": true,
     "isolatedModules": true,
     "jsx": "preserve",
@@ -10,6 +11,9 @@
     "module": "commonjs",
     "noEmit": true,
     "noUnusedLocals": false,
+    "paths": {
+      "*": ["./gatsby-ts/types/*"]
+    },
     "strict": true,
     "target": "esnext"
   },
